move around packets
